feat: Add defence stalker position for Protoss bot
This commit adds a new attribute `defence_stalker_position` to the Protoss bot class. This position is used to assign stalkers to defend a specific location on the map. The position is determined based on the starting side of the bot and the map center. This enhancement improves the defensive capabilities of the bot.

Note: The `defence_stalker_position` logic is currently not working as expected.